CR2 Battery
The CR2 battery, which is found in cameras,weapon-mounted lights, golf rangefinders, EDC flashlights, and security devices, is revered around the world. A standard cylindrical lithium battery is referred to as a CR2. This is the type of battery used in cameras (such as the CR2). This is due to their small size, which makes them ideal for electronic devices on the go. A CR2 lithium battery's shelf life is determined by its model and manufacturer. The Energizer brand, for example, puts more strain on its batteries than other brands, ensuring comparative endurance under extreme conditions.
https://www.apogeeweb.net/electron/cr2-battery-basics-replacement-and-safety-precautions-faq.html